requirements.txt
flask>=3.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
gunicorn>=21.2.0
Jinja2>=3.1.4
click>=8.1.7
waitress>=3.0.0
